% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gym_env.R
\name{build_gym_env}
\alias{build_gym_env}
\title{Build a new gym environment}
\usage{
build_gym_env(...)
}
\arguments{
\item{...}{Members to add to the new gym environment. This must contain at least a 'func', 'action_space' and 'observation_space' member}
}
\description{
Take a function intended for training with reinforcement learning and create a valid gym environment object.
}
\section{Required members}{

A new gym environment requires the following key members:
\describe{
  \item{func}{The func function you want to learn parameters for}
  \item{action_space}{A vector of possible actions for the func}
  \item{observation_space}{A vector of possible observations from the func function}
}
}

\section{Optional members}{

There are also the following optional members:
\describe{
   \item{random}{A number between 0 and 1 dictating the level of randomness applied when running the func function. A lower number means it will
   return the intended result less often (exploring the total observation space more often)}
}
}

\examples{
\dontrun{
build_gym_env(func = nchain_function,
  action_space = c(1, 2),
  observation_space = 1:5),
  random = 0.5
 }
}
