% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gym.R
\docType{data}
\name{gym_env}
\alias{gym_env}
\title{Build a gym environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
gym_env
}
\description{
Build a gym environment
}
\section{Creating the gym}{

A gym environment is an R6 object, that can be created with \code{gym_env$new()}. It has the following arguments:
\describe{
  \item{env}{The environment we want to perform reinforcement learning on. This is a function object that determines rewards and new states
  based on input states and actions.}
  \item{length}{The number of actions within the environment until it is considered to be done. By default this is 1,000 actions.}
}
}

\section{Working out at the gym}{

There are three key functions to use with a gym class.
\describe{
  \item{\code{gym_env$step(a)}}{This function allows us to step through the environment we have defined above. We supply an action, and gym then
  runs this action through the environment}
  \item{\code{gym_env$reset()}}{When we don't want to worry about recreating a gym object, we can just reset the object back to the state of
  creation. This clears all cumulative rewards etc.}
  \item{\code{gym_env$get()}}{This function allows us to extract the private values within the R6 class.}
}
}

\keyword{datasets}
