% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gym_env.R
\name{gym_new}
\alias{gym_new}
\title{Create a new gym environment}
\usage{
gym_new(...)
}
\arguments{
\item{...}{Members to add to the gym environment. This must contain at least a 'game', 'action_space' and 'observation_space' member}
}
\description{
Create a new gym environment
}
