% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gymr.R
\docType{package}
\name{gymr-package}
\alias{gymr}
\alias{gymr-package}
\title{gymr: Build Environments to Test Reinforcement Learning Algorithms}
\description{
A package inspired from Open AI's gym in Python. The package is designed to provide a lightweight way of building 
  and training reinforcement learning algorithms inside of R.
}
\author{
\strong{Maintainer}: Liam Gilbey \email{liam.gilbey@gmail.com}

}
\keyword{internal}
