% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gymr.R
\docType{package}
\name{gymr-package}
\alias{gymr}
\alias{gymr-package}
\title{gymr: Build Environments to Test Reinforcement Learning Algorithms}
\description{
Inspired from Open AI's gym in Python.
}
\author{
\strong{Maintainer}: Liam Gilbey \email{liam.gilbey@gmail.com}

}
\keyword{internal}
